backend/requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# AI/ML
google-generativeai==0.3.0
requests==2.31.0

# Domain Analysis
whois==0.9.7

# Async
aiohttp==3.9.1

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1

# Optional: Production
gunicorn==21.2.0

# Logging
colorama==0.4.6
